Fix Windows build

+++ arangod/RocksDBEngine/RocksDBRecoveryManager.cpp
@@ -430,7 +430,7 @@ class WBReader final : public rocksdb::WriteBatch::Handler {
       }
 
       uint64_t currentCount = coll->meta().numberDocuments();
-      coll->meta().adjustNumberDocumentsInRecovery(_currentSequence, -currentCount, 0);
+      coll->meta().adjustNumberDocumentsInRecovery(_currentSequence, -static_cast<int64_t>(currentCount), 0);
       for (std::shared_ptr<arangodb::Index> const& idx : coll->getIndexes()) {
         RocksDBIndex* ridx = static_cast<RocksDBIndex*>(idx.get());
         RocksDBCuckooIndexEstimator<uint64_t>* est = ridx->estimator();
